+++ hdiv-core/src/main/java/org/hdiv/filter/IValidationHelper.java
@@ -79,4 +79,13 @@ public interface IValidationHelper extends StateRestorer {
 	 */
 	List<ValidatorError> findCustomErrors(final Throwable t, String target);
 
+	/**
+	 * Check whether all the errors are legal
+	 * @param errors
+	 * @return
+	 */
+	boolean areErrorsLegal(List<ValidatorError> errors);
+
+	boolean processEditableValidationErrors(final RequestContextHolder request, final List<ValidatorError> errors);
+
 }