+++ hdiv-core/src/test/java/org/hdiv/filter/ValidatorHelperTest.java
@@ -679,6 +679,58 @@ public class ValidatorHelperTest extends AbstractHDIVTestCase {
 		assertTrue(result);
 	}
 
+	/**
+	 * Test for cookies integrity.
+	 */
+	public void testValidateClientCookies() {
+
+		MockHttpServletRequest request = getMockRequest();
+
+		dataComposer.beginRequest(Method.GET, targetName);
+		dataComposer.compose("param1", "value1", false);
+		String pageState = dataComposer.endRequest();
+		assertNotNull(pageState);
+		addParameter(pageState);
+
+		dataComposer.endPage();
+
+		// Modify cookie value on client
+		request.setCookies(new Cookie[] { new Cookie("name", "changedValue") });
+
+		boolean result = helper.validate(context).isValid();
+		assertTrue(result);
+	}
+
+	/**
+	 * Test for cookies integrity.
+	 */
+	public void testValidateClientAndServerCookies() {
+
+		MockHttpServletRequest request = getMockRequest();
+
+		responseWrapper.addCookie(new Cookie("name", "value"));
+
+		dataComposer.beginRequest(Method.GET, targetName);
+		dataComposer.compose("param1", "value1", false);
+		String pageState = dataComposer.endRequest();
+		assertNotNull(pageState);
+		addParameter(pageState);
+
+		dataComposer.endPage();
+
+		// Add new cookie
+		request.setCookies(new Cookie[] { new Cookie("name2", "value2") });
+
+		boolean result = helper.validate(context).isValid();
+		assertTrue(result);
+
+		// modify cookie value
+		request.setCookies(new Cookie[] { new Cookie("name", "changedValue") });
+
+		result = helper.validate(context).isValid();
+		assertFalse(result);
+	}
+
 	public void testValidateWhitespace() {
 
 		MockHttpServletRequest request = getMockRequest();
@@ -1255,5 +1307,60 @@ public class ValidatorHelperTest extends AbstractHDIVTestCase {
 		// fix = valueññ
 
 		assertEquals(value, fix);
+	}
+
+	public void testSameActionParamAndField() {
+
+		MockHttpServletRequest request = getMockRequest();
+
+		dataComposer.startPage();
+		dataComposer.beginRequest(Method.POST, targetName);
+		dataComposer.composeParams("a=A", Method.POST, Constants.ENCODING_UTF_8);
+		dataComposer.composeFormField("a", "A", false, "hidden", true);
+		String pageState = dataComposer.endRequest();
+		dataComposer.endPage();
+
+		request.addParameter(hdivParameter, pageState);
+		request.addParameter("a", "0");
+
+		ValidatorHelperResult result = helper.validate(context);
+		assertTrue(result.isValid());
+	}
+
+	public void testSameActionParamAndFieldNoConfidentiality() {
+
+		getConfig().setConfidentiality(false);
+		MockHttpServletRequest request = getMockRequest();
+
+		dataComposer.startPage();
+		dataComposer.beginRequest(Method.POST, targetName);
+		dataComposer.composeParams("a=1", Method.POST, Constants.ENCODING_UTF_8);
+		dataComposer.composeFormField("a", "1", false, "hidden", true);
+		String pageState = dataComposer.endRequest();
+		dataComposer.endPage();
+
+		request.addParameter(hdivParameter, pageState);
+		request.addParameter("a", "1");
+
+		ValidatorHelperResult result = helper.validate(context);
+		assertTrue(result.isValid());
+	}
+
+	public void testValidateMissingParameters() {
+
+		getConfig().setConfidentiality(false);
+		MockHttpServletRequest request = getMockRequest();
+
+		dataComposer.startPage();
+		dataComposer.beginRequest(Method.GET, targetName);
+		dataComposer.composeParams("a=1&clientGeneratedParam=2", Method.POST, Constants.ENCODING_UTF_8);
+		String pageState = dataComposer.endRequest();
+		dataComposer.endPage();
+
+		request.addParameter(hdivParameter, pageState);
+		request.addParameter("a", "1");
+
+		ValidatorHelperResult result = helper.validate(context);
+		assertTrue(result.isValid());
 	}
 }